Clean up the old edit-form mode, and rename edit-geometry to just edit
