+++ src/lib/draw/EditMode.svelte
