+++ src/lib/draw/index.ts
@@ -1,5 +1,4 @@
 export { default as BoundaryLayer } from "./BoundaryLayer.svelte";
-export { default as EditGeometryMode } from "./EditGeometryMode.svelte";
 export { default as ImageLayer } from "./image/ImageLayer.svelte";
 export { default as HoverLayer } from "./HoverLayer.svelte";
 export { default as InterventionLayer } from "./InterventionLayer.svelte";
